# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FleetManager API
  description: API для системы управления автопарком
  version: 1.0.0
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Car:
      type: object
      properties:
        _id: { type: string }
        brand: { type: string }
        model: { type: string }
        plate_number: { type: string }
        vin: { type: string }
        year: { type: integer }
        status: { type: string, enum: [active, repair, reserve] }
        mileage: { type: number }
        assigned_to: { type: string, nullable: true }
        equipment: { type: array, items: { $ref: '#/components/schemas/Equipment' } }
        photos: { type: array, items: { type: string } }
        documents: { type: array, items: { type: string } }
    Equipment:
      type: object
      properties:
        name: { type: string }
        serial_number: { type: string }
        installed_date: { type: string }
    Employee:
      type: object
      properties:
        _id: { type: string }
        full_name: { type: string }
        position: { type: string }
        department: { type: string }
        phone: { type: string }
        status: { type: string, enum: [active, inactive] }
        hire_date: { type: string }
        card_id: { type: string, nullable: true }
        assigned_cars: { type: array, items: { type: string } }
    Card:
      type: object
      properties:
        _id: { type: string }
        card_number: { type: string }
        type: { type: string, enum: [fuel, service] }
        balance: { type: number }
        limit: { type: number }
        status: { type: string, enum: [active, blocked, expired] }
        assigned_to: { type: string, nullable: true }
        assigned_car: { type: string, nullable: true }
        expiry_date: { type: string, nullable: true }
    Transaction:
      type: object
      properties:
        _id: { type: string }
        card_id: { type: string }
        employee_id: { type: string }
        car_id: { type: string }
        amount: { type: number }
        fuel_type: { type: string }
        volume: { type: number }
        location: { type: string }
        odometer: { type: number }
        date: { type: string }
        status: { type: string, enum: [completed, pending, cancelled] }
...
    Paginated:
      type: object
      properties:
        data: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
paths:
...
  /cars:
    get:
      summary: Список автомобилей
      tags: [Cars]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: page, in: query, schema: { type: integer } }, { name: limit, in: query, schema: { type: integer } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } } }
    post:
      summary: Создать автомобиль
      tags: [Cars]
      security: [{ bearerAuth: [] }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Car' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Car' } } } }
        '400': { description: Validation error }
...
  /employees:
    get:
      summary: Список сотрудников
      tags: [Employees]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: page, in: query }, { name: limit, in: query }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } } }
    post:
      summary: Создать сотрудника
      tags: [Employees]
      security: [{ bearerAuth: [] }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Employee' } } } }
      responses:
        '201': { description: Created }
...
  /cards:
    get:
      summary: Список карт
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: page, in: query }, { name: limit, in: query }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } } }
    post:
      summary: Создать карту
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Card' } } } }
      responses:
        '201': { description: Created }
...
  /transactions:
    get:
      summary: Список транзакций
      tags: [Transactions]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: page, in: query }, { name: limit, in: query }, { name: card_id, in: query }, { name: employee_id, in: query }, { name: dateFrom, in: query }, { name: dateTo, in: query }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } } }
    post:
      summary: Создать транзакцию
      tags: [Transactions]
      security: [{ bearerAuth: [] }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
      responses:
        '201': { description: Created }
...
  /maintenance:
    get:
      summary: История ТО
      tags: [Maintenance]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: car_id, in: query }]
      responses:
        '200': { description: OK }
    post:
      summary: Создать запись ТО
      tags: [Maintenance]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Created }
...